Separate the recorder into its own file.
