+++ lib/trimHeaders.js
@@ -1,0 +1,6 @@
+module.exports = function trimHeaders(message) {
+  delete message.headers['Content-Length'];
+  delete message.headers['Transfer-Encoding'];
+  delete message.headers.Connection;
+  delete message.headers.Date;
+};
