+++ lib/unexpectedMitm.js
@@ -1,44 +1,19 @@
 /* global setImmediate, process, after, console */
 const messy = require('messy');
-const consumeReadableStream = require('./consumeReadableStream');
-const createMitm = require('mitm-papandreou');
-const createSerializedRequestHandler = require('./createSerializedRequestHandler');
-const _ = require('underscore');
-const http = require('http');
-const https = require('https');
 const formatHeaderObj = require('./formatHeaderObj');
 const fs = require('fs');
 const path = require('path');
 const memoizeSync = require('memoizesync');
 const callsite = require('callsite');
 const detectIndent = require('detect-indent');
-const metadataPropertyNames = messy.HttpRequest.metadataPropertyNames.concat(
-  'rejectUnauthorized'
-);
 
+const performRequest = require('./performRequest');
+const trimHeaders = require('./trimHeaders');
 const UnexpectedMitmMocker = require('./UnexpectedMitmMocker');
+const UnexpectedMitmRecorder = require('./UnexpectedMitmRecorder');
 
 function checkEnvFlag(varName) {
   return process.env[varName] === 'true';
-}
-
-function isTextualBody(message, content) {
-  return message.hasTextualContentType && bufferCanBeInterpretedAsUtf8(content);
-}
-
-function bufferCanBeInterpretedAsUtf8(buffer) {
-  // Hack: Since Buffer.prototype.toString('utf-8') is very forgiving, convert the buffer to a string
-  // with percent-encoded octets, then see if decodeURIComponent accepts it.
-  try {
-    decodeURIComponent(
-      Array.prototype.map
-        .call(buffer, octet => `%${octet < 16 ? '0' : ''}${octet.toString(16)}`)
-        .join('')
-    );
-  } catch (e) {
-    return false;
-  }
-  return true;
 }
 
 function lineNumberToIndex(string, lineNumber) {
@@ -53,79 +28,6 @@ function lineNumberToIndex(string, lineNumber) {
   }
 
   return startOfLineIndex;
-}
-
-function trimHeaders(message) {
-  delete message.headers['Content-Length'];
-  delete message.headers['Transfer-Encoding'];
-  delete message.headers.Connection;
-  delete message.headers.Date;
-}
-
-function trimMessage(message) {
-  if (typeof message.body !== 'undefined') {
-    if (message.body.length === 0) {
-      delete message.body;
-    } else if (
-      isTextualBody(
-        new messy.Message({
-          headers: { 'Content-Type': message.headers['Content-Type'] }
-        }),
-        message.body
-      )
-    ) {
-      message.body = message.body.toString('utf-8');
-    }
-    if (
-      /^application\/json(?:;|$)|\+json\b/.test(
-        message.headers['Content-Type']
-      ) &&
-      /^\s*[[{]/.test(message.body)
-    ) {
-      try {
-        message.body = JSON.parse(message.body);
-        if (message.headers['Content-Type'] === 'application/json') {
-          delete message.headers['Content-Type'];
-        }
-      } catch (e) {}
-    }
-  }
-  if (message.statusCode === 200) {
-    delete message.statusCode;
-  }
-  if (message.headers) {
-    trimHeaders(message);
-    if (Object.keys(message.headers).length === 0) {
-      delete message.headers;
-    }
-  }
-  if (message.url && message.method) {
-    message.url = `${message.method} ${message.url}`;
-    delete message.method;
-  }
-  // Remove properties with an undefined value. Prevents things
-  // like rejectUnauthorized:true from showing up in recordings:
-  Object.keys(message).forEach(key => {
-    if (typeof message[key] === 'undefined') {
-      delete message[key];
-    }
-  });
-  if (Object.keys(message).length === 1) {
-    if (typeof message.url === 'string') {
-      return message.url;
-    }
-    if (typeof message.statusCode === 'number') {
-      return message.statusCode;
-    }
-  }
-  return message;
-}
-
-function trimRecordedExchange(recordedExchange) {
-  return {
-    request: trimMessage(recordedExchange.request),
-    response: trimMessage(recordedExchange.response)
-  };
 }
 
 function determineCallsite(stack) {
@@ -337,42 +239,6 @@ module.exports = {
       }
     }
 
-    function performRequest(requestResult) {
-      return expect.promise((resolve, reject) => {
-        (requestResult.encrypted ? https : http)
-          .request(
-            Object.assign(
-              {
-                headers: requestResult.headers,
-                method: requestResult.method,
-                host: requestResult.host,
-                port: requestResult.port,
-                path: requestResult.path
-              },
-              requestResult.metadata
-            )
-          )
-          .on('response', response => {
-            consumeReadableStream(response)
-              .catch(reject)
-              .then(result => {
-                if (result.error) {
-                  // TODO: Consider adding support for recording this (the upstream response erroring out while we're recording it)
-                  return reject(result.error);
-                }
-
-                resolve({
-                  statusCode: response.statusCode,
-                  headers: response.headers,
-                  body: result.body
-                });
-              });
-          })
-          .on('error', reject)
-          .end(requestResult.body);
-      });
-    }
-
     function applyInjections() {
       Object.keys(injectionsBySourceFileName).forEach(sourceFileName => {
         const injections = injectionsBySourceFileName[sourceFileName];
@@ -462,141 +328,6 @@ module.exports = {
       };
     }
 
-    function executeMitm(expect, subject) {
-      const mitm = createMitm();
-
-      let recordedExchanges = [];
-
-      return expect
-        .promise((resolve, reject) => {
-          let bypassNextConnect = false;
-
-          mitm
-            .on('connect', (socket, opts) => {
-              if (bypassNextConnect) {
-                socket.bypass();
-                bypassNextConnect = false;
-              }
-            })
-            .on(
-              'request',
-              createSerializedRequestHandler((req, res) => {
-                const clientSocket = req.connection._mitm.client;
-                const clientSocketOptions = req.connection._mitm.opts;
-                const metadata = Object.assign(
-                  {},
-                  _.pick(
-                    clientSocketOptions.agent &&
-                      clientSocketOptions.agent.options,
-                    metadataPropertyNames
-                  ),
-                  _.pick(clientSocketOptions, metadataPropertyNames)
-                );
-
-                const recordedExchange = {
-                  request: Object.assign(
-                    {
-                      url: `${req.method} ${req.url}`,
-                      headers: formatHeaderObj(req.headers)
-                    },
-                    metadata
-                  ),
-                  response: {}
-                };
-                recordedExchanges.push(recordedExchange);
-                consumeReadableStream(req)
-                  .catch(reject)
-                  .then(result => {
-                    if (result.error) {
-                      // TODO: Consider adding support for recording this (the request erroring out while we're recording it)
-                      return reject(result.error);
-                    }
-                    recordedExchange.request.body = result.body;
-                    bypassNextConnect = true;
-                    const matchHostHeader =
-                      req.headers.host &&
-                      req.headers.host.match(/^([^:]*)(?::(\d+))?/);
-
-                    let host;
-
-                    let port;
-
-                    // https://github.com/moll/node-mitm/issues/14
-                    if (matchHostHeader) {
-                      if (matchHostHeader[1]) {
-                        host = matchHostHeader[1];
-                      }
-                      if (matchHostHeader[2]) {
-                        port = parseInt(matchHostHeader[2], 10);
-                      }
-                    }
-                    if (!host) {
-                      return reject(
-                        new Error(
-                          `unexpected-mitm recording mode: Could not determine the host name from Host header: ${
-                            req.headers.host
-                          }`
-                        )
-                      );
-                    }
-
-                    performRequest({
-                      encrypted: req.socket.encrypted,
-                      headers: req.headers,
-                      method: req.method,
-                      host,
-                      // default the port to HTTP values if not set
-                      port: port || (req.socket.encrypted ? 443 : 80),
-                      path: req.url,
-                      body: result.body,
-                      metadata
-                    })
-                      .then(responseResult => {
-                        recordedExchange.response.statusCode =
-                          responseResult.statusCode;
-                        recordedExchange.response.headers = formatHeaderObj(
-                          responseResult.headers
-                        );
-                        recordedExchange.response.body = responseResult.body;
-
-                        setImmediate(() => {
-                          res.statusCode = responseResult.statusCode;
-                          Object.keys(responseResult.headers).forEach(
-                            headerName => {
-                              res.setHeader(
-                                headerName,
-                                responseResult.headers[headerName]
-                              );
-                            }
-                          );
-                          res.end(recordedExchange.response.body);
-                        });
-                      })
-                      .catch(err => {
-                        recordedExchange.response = err;
-                        clientSocket.emit('error', err);
-                      });
-                  });
-              })
-            );
-
-          expect
-            .promise(() => expect.shift())
-            .catch(reject)
-            .then(value => {
-              recordedExchanges = recordedExchanges.map(trimRecordedExchange);
-              if (recordedExchanges.length === 1) {
-                recordedExchanges = recordedExchanges[0];
-              }
-
-              resolve([value, recordedExchanges]);
-            });
-        })
-        .finally(() => {
-          mitm.disable();
-        });
-    }
-
     let afterBlockRegistered = false;
 
     expect
@@ -612,8 +343,11 @@ module.exports = {
             afterBlockRegistered = true;
           }
 
-          return executeMitm(expect, subject).then(
-            ([value, recordedExchanges]) => {
+          const recorder = new UnexpectedMitmRecorder();
+
+          return expect
+            .promise(() => recorder.record(() => expect.shift()))
+            .then(([fulfilmentValue, recordedExchanges]) => {
               if (injectIntoTest) {
                 const injectionCallsite = determineCallsite(stack);
                 if (injectionCallsite) {
@@ -621,12 +355,11 @@ module.exports = {
                 }
               }
               if (expect.flags['with extra info']) {
-                return [value, recordedExchanges];
+                return [fulfilmentValue, recordedExchanges];
               } else {
-                return value;
+                return fulfilmentValue;
               }
-            }
-          );
+            });
         }
       )
       .exportAssertion(
@@ -644,8 +377,11 @@ module.exports = {
           }
 
           if (checkEnvFlag('UNEXPECTED_MITM_WRITE')) {
-            return executeMitm(expect, subject).then(
-              ([fulfilmentValue, recordedExchanges]) => {
+            const recorder = new UnexpectedMitmRecorder();
+
+            return expect
+              .promise(() => recorder.record(() => expect.shift()))
+              .then(([fulfilmentValue, recordedExchanges]) => {
                 const output = `module.exports = ${stringify(
                   recordedExchanges,
                   4
@@ -658,8 +394,7 @@ module.exports = {
                 } else {
                   return recordedExchanges;
                 }
-              }
-            );
+              });
           }
 
           return expect
